Move transient contract remover in transaction conversion
