+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/index/TransactionConversion.scala
@@ -5,7 +5,6 @@ package com.digitalasset.platform.index
 
 import com.digitalasset.api.util.TimestampConversion
 import com.digitalasset.daml.lf.data.Ref
-import com.digitalasset.daml.lf.data.Relation.Relation
 import com.digitalasset.daml.lf.engine
 import com.digitalasset.daml.lf.engine.Blinding
 import com.digitalasset.daml.lf.transaction.Node.{GenNode, NodeCreate, NodeExercises}
@@ -13,7 +12,8 @@ import com.digitalasset.daml.lf.value.Value.AbsoluteContractId
 import com.digitalasset.daml.lf.value.{Value => Lf}
 import com.digitalasset.ledger.EventId
 import com.digitalasset.ledger.api.domain
-import com.digitalasset.ledger.api.v1.event.{CreatedEvent, Event, ExercisedEvent}
+import com.digitalasset.ledger.api.v1.event.Event.Event.{Archived, Created, Empty}
+import com.digitalasset.ledger.api.v1.event.{ArchivedEvent, CreatedEvent, Event, ExercisedEvent}
 import com.digitalasset.ledger.api.v1.transaction.{Transaction, TransactionTree, TreeEvent}
 import com.digitalasset.platform.api.v1.event.EventOps.TreeEventOps
 import com.digitalasset.platform.common.PlatformTypes.{CreateEvent, ExerciseEvent}
@@ -21,13 +21,14 @@ import com.digitalasset.platform.events.EventIdFormatter
 import com.digitalasset.platform.participant.util.LfEngineToApi
 import com.digitalasset.platform.participant.util.LfEngineToApi.{
   assertOrRuntimeEx,
-  lfNodeCreateToCreatedEvent,
-  lfNodeExerciseToArchivedEvent
+  lfNodeCreateToFlatApiCreated,
+  lfNodeExerciseToFlatApiArchived
 }
-import com.digitalasset.platform.server.services.transaction.TransientContractRemover
 import com.digitalasset.platform.store.entries.LedgerEntry
+import com.digitalasset.platform.api.v1.event.EventOps.EventOps
 
 import scala.annotation.tailrec
+import scala.collection.{breakOut, mutable}
 
 object TransactionConversion {
 
@@ -35,30 +36,80 @@ object TransactionConversion {
   private type Create = NodeCreate.WithTxValue[Lf.AbsoluteContractId]
   private type Exercise = NodeExercises.WithTxValue[EventId, Lf.AbsoluteContractId]
 
-  private def flatEvent(
-      disclosure: Relation[EventId, Ref.Party],
-      verbose: Boolean): PartialFunction[(EventId, Node), Event] = {
+  private def flatEvent(verbose: Boolean): PartialFunction[(EventId, Node), Event] = {
     case (eventId, node: Create) =>
       assertOrRuntimeEx(
         failureContext = "converting a create node to a created event",
-        lfNodeCreateToCreatedEvent(
-          verbose = verbose,
-          witnessParties = node.stakeholders,
-          eventId = eventId,
-          node = node,
-        )
+        lfNodeCreateToFlatApiCreated(verbose, eventId, node)
       )
     case (eventId, node: Exercise) if node.consuming =>
       assertOrRuntimeEx(
         failureContext = "converting a consuming exercise node to an archived event",
-        lfNodeExerciseToArchivedEvent(
-          witnessParties = node.stakeholders,
-          eventId = eventId,
-          node = node,
-        )
+        lfNodeExerciseToFlatApiArchived(eventId, node)
       )
-
   }
+
+  /**
+    * Cancels out witnesses on creates and archives that are about the same contract.
+    * If no witnesses remain on either, the node is removed.
+    *
+    * @param nodes Must be sorted by event index.
+    * @throws IllegalArgumentException if the argument is not sorted properly.
+    */
+  private[index] def removeTransients(nodes: Vector[Event]): Vector[Event] = {
+
+    val resultBuilder = new Array[Option[Event]](nodes.size)
+    val creationByContractId = new mutable.HashMap[String, (Int, Event)]()
+
+    nodes.zipWithIndex.foreach {
+      case (event, indexInList) =>
+        // Each call adds a new (possibly null) element to resultBuilder, and may update items previously added
+        updateResultBuilder(resultBuilder, creationByContractId, event, indexInList)
+    }
+
+    resultBuilder.collect {
+      case Some(v) if v.witnessParties.nonEmpty => v
+    }(breakOut)
+  }
+
+  /**
+    * Update resultBuilder given the next event.
+    * This will insert a new element and possibly update a previous one.
+    */
+  private def updateResultBuilder(
+      resultBuilder: Array[Option[Event]],
+      creationByContractId: mutable.HashMap[String, (Int, Event)],
+      event: Event,
+      indexInList: Int
+  ): Unit =
+    event match {
+      case createdEvent @ Event(
+            Created(CreatedEvent(_, contractId, _, _, witnessParties, _, _, _, _))) =>
+        if (witnessParties.nonEmpty) {
+          resultBuilder.update(indexInList, Some(event))
+          val _ = creationByContractId.put(contractId, indexInList -> createdEvent)
+        }
+      case archivedEvent @ Event(Archived(ArchivedEvent(_, contractId, _, witnessParties))) =>
+        if (witnessParties.nonEmpty) {
+          creationByContractId
+            .get(contractId)
+            .fold[Unit] {
+              // No matching create for this archive. Insert as is.
+              resultBuilder.update(indexInList, Some(event))
+            } {
+              case (createdEventIndex, createdEvent) =>
+                // Defensive code to ensure that the set of parties the events are disclosed to are not different.
+                if (witnessParties.toSet != createdEvent.getCreated.witnessParties.toSet)
+                  throw new IllegalArgumentException(
+                    s"Created and Archived event stakeholders are different in $createdEvent, $archivedEvent")
+
+                resultBuilder.update(createdEventIndex, None)
+                resultBuilder.update(indexInList, None)
+            }
+        }
+      case Event(Empty) =>
+        throw new IllegalArgumentException("Empty event")
+    }
 
   def ledgerEntryToFlatTransaction(
       offset: domain.LedgerOffset.Absolute,
@@ -67,8 +118,8 @@ object TransactionConversion {
       verbose: Boolean,
   ): Option[Transaction] = {
 
-    val flatEvents = entry.transaction.collect(flatEvent(entry.explicitDisclosure, verbose))
-    val withoutTransientContracts = TransientContractRemover.removeTransients(flatEvents)
+    val flatEvents = entry.transaction.collect(flatEvent(verbose))
+    val withoutTransientContracts = removeTransients(flatEvents)
     val withFilteredWitnesses = withoutTransientContracts.flatMap(EventFilter(_)(filter).toList)
 
     val commandId =
